+++ src/main/java/com/uqpay/sdk/banking/model/VirtualAccountBankDetailStatus.java
@@ -1,0 +1,6 @@
+package com.uqpay.sdk.banking.model;
+
+public enum VirtualAccountBankDetailStatus {
+    ACTIVE,
+    CLOSED
+}