+++ src/main/java/com/uqpay/sdk/banking/model/VirtualAccountClearingSystem.java
@@ -1,0 +1,29 @@
+package com.uqpay.sdk.banking.model;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class VirtualAccountClearingSystem {
+    @JsonProperty("type")
+    private String type;
+
+    @JsonProperty("value")
+    private String value;
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    public String getValue() {
+        return value;
+    }
+
+    public void setValue(String value) {
+        this.value = value;
+    }
+}